changed tissuepct_value to float

+++ Tiling.py
@@ -15,7 +15,7 @@ parser.add_argument('--target_path', type=str)
 parser.add_argument('--workers',type=int,default=8)
 parser.add_argument('--tilesize',type=int,default=256)
 parser.add_argument('--stride',type=int,default=256)
-parser.add_argument('--tissuepct_value',type=int,default=0.7)
+parser.add_argument('--tissuepct_value',type=float,default=0.7)
 parser.add_argument('--magnification',type=str, default = 'multi')
 
 def tiling(svspath,patientid,targetpath,tilesize=256,stride=256,tissuepct_value=0.7,magnification='multi'):
